Bugfix:  error in Version string 'Unknown': version number does not start with digit
resolves: https://github.com/Hoglet/TimeIT/issues/3

Version number is taken from git (fallback: debian version number) and is generated on build. Version number for CPACK (debian and red hat packages) are stored at configuration time and is not updated at builds.

This is insuring that the version number is created, even though it might be outdated as subsequent builds are not causing a reconfiguration.